alzheimer_detection/requirements.txt
Django
django-debug-toolbar
djangorestframework
gunicorn
Keras
Keras-Applications
Keras-Preprocessing
numpy
pandas
pytest
tensorflow
psycopg2
requests
setuptools